Reverted deletion of the Date initialization in Diaper.
Added Submit button to Feeding Overlay with possible function.
onFeedingSubmit includes the getvalues for the name, Oz, type.

Added a method to User.Java for the feeding update from
the submit click on MainActivity.

setUserFeeding checks babyname and will update that
babies feeding. //May need to add a check for type of feeding.